+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.1.3](https://github.com/bitrix24/b24ui/compare/v2.1.2...v2.1.3) (2025-11-xx)
+
+* chore(deps): update `nuxt-component-meta`
+
 ## [2.1.2](https://github.com/bitrix24/b24ui/compare/v2.1.1...v2.1.2) (2025-11-13)
 
 ### Features